Move font layout logic into text component.
